Move the serializer and xml wrapper headers out of public/internal and into private folders within src.
