+++ sdk/tables/azure-data-tables/CMakeLists.txt
@@ -50,8 +50,6 @@ set(
     inc/azure/data/tables/internal/policies/shared_key_lite_policy.hpp
     inc/azure/data/tables/internal/policies/tenant_bearer_token_policy.hpp
     inc/azure/data/tables/internal/policies/timeout_policy.hpp
-    inc/azure/data/tables/internal/serializers.hpp
-    inc/azure/data/tables/internal/xml_wrapper.hpp
     inc/azure/data/tables/models.hpp
     inc/azure/data/tables/rtti.hpp
     inc/azure/data/tables.hpp
@@ -69,8 +67,10 @@ set(
     src/policies/tenant_bearer_token_policy.cpp
     src/policies/timeout_policy.cpp
     src/private/hmacsha256.hpp
-    src/private/url_encode.hpp
     src/private/package_version.hpp
+    src/private/serializers.hpp
+    src/private/url_encode.hpp
+    src/private/xml_wrapper.hpp
     src/serializers.cpp
     src/tables_clients.cpp
     src/tables_sas_builder.cpp
