+++ sdk/tables/azure-data-tables/src/private/serializers.hpp
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 #pragma once
-#include "azure/data/tables/internal/xml_wrapper.hpp"
+#include "xml_wrapper.hpp"
 #include "azure/data/tables/models.hpp"
 
 #include <azure/core/internal/json/json.hpp>
