+++ sdk/tables/azure-data-tables/src/private/xml_wrapper.hpp
