+++ sdk/tables/azure-data-tables/src/tables_clients.cpp
@@ -7,7 +7,7 @@
 #include "azure/data/tables/internal/policies/shared_key_lite_policy.hpp"
 #include "azure/data/tables/internal/policies/tenant_bearer_token_policy.hpp"
 #include "azure/data/tables/internal/policies/timeout_policy.hpp"
-#include "azure/data/tables/internal/serializers.hpp"
+#include "private/serializers.hpp"
 
 #include <sstream>
 #include <string>
