+++ sdk/tables/azure-data-tables/test/ut/serializers_test.hpp
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#include "azure/data/tables/internal/serializers.hpp"
+#include "../src/private/serializers.hpp"
 #include "azure/data/tables/tables_clients.hpp"
 #include "test/ut/test_base.hpp"
